On branch rework  Your branch is up-to-date with 'origin/rework'.  Changes to be committed: 	modified:   Assign_Management/__pycache__/models.cpython-36.pyc 	modified:   Assign_Management/__pycache__/views.cpython-36.pyc 	modified:   Assign_Management/models.py 	modified:   Assign_Management/views.py 	modified:   Class_Management/Template/ShowQuizListStudent.html 	modified:   Class_Management/Template/ShowQuizStudent.html 	modified:   Class_Management/Template/ShowStudent.html 	modified:   Class_Management/__pycache__/models.cpython-36.pyc 	modified:   Class_Management/__pycache__/views.cpython-36.pyc 	modified:   Class_Management/models.py 	modified:   Class_Management/views.py 	modified:   db.sqlite3 	deleted:    media/59340500018_coded_eee_FRA141_mhReGjc.py 	deleted:    media/59340500018_coded_eee_FRA141_n4QRHvy.py 	deleted:    media/59340500018_coded_eee_FRA141_sGtj8KA.py 	deleted:    media/59340500018_coded_eee_FRA141_vYLzSyw.py 	deleted:    media/59340500018_coded_eee_FRA141_vaJYIuG.py 	deleted:    media/59340500018_coded_eee_FRA141_xZ5QG0Z.py 	deleted:    media/admin_adfsf_script.py

+++ Class_Management/views.py
@@ -376,8 +376,6 @@ def StudentScoreInfo(request,classroom,username):
         u_id = request.session['u_id']
         var = request.user.username
         print(u_id[0])
-        if username != var and not request.user.is_admin:
-            return HttpResponseRedirect("/ClassRoom/"+request.session["classroom"])
         try:
             print("try")
             score = QuizScore.objects.filter(studentId=User.objects.get(username=u_id[0]), classroom=ClassRoom.objects.get(className=classroom))
@@ -407,10 +405,11 @@ def StudentQuizListInfo(request,classroom,username,quiz_id):
     if not request.user.is_authenticated:
         return HttpResponseRedirect('/LogOut')
     else:
-        if request.method == 'POST':
-            title_list = request.POST.getlist("cb")
-            #print(title_list)
-            Upload.objects.filter(title__in=title_list).delete()
+        if request.user.username == username or request.user.is_admin:
+            if request.method == 'POST':
+                title_list = request.POST.getlist("cb")
+                #print(title_list)
+                Upload.objects.filter(title__in=title_list).delete()
         file_list = Upload.objects.filter(user=User.objects.get(username=username),
                                           quiz=Quiz.objects.get(pk=quiz_id),
                                           classroom=Quiz.objects.get(pk=quiz_id).classroom
@@ -435,7 +434,9 @@ def StudentQuizInfo(request,classroom,username,quiz_id,title):
     if not request.user.is_authenticated:
         return HttpResponseRedirect('/LogOut')
 
-    elif request.method == 'POST' and request.user.is_admin:
+    elif request.method == 'POST' and request.user.is_admin or request.method == 'POST' and request.user.username == username:
+        if username != request.user.username and not request.user.is_admin:
+            return HttpResponseRedirect("/ClassRoom/Home")
         score_pointer = QuizScore.objects.get(quizId=Quiz.objects.get(pk=quiz_id),
                                               studentId=User.objects.get(username=username)
                                               )
@@ -449,7 +450,7 @@ def StudentQuizInfo(request,classroom,username,quiz_id,title):
 
     else:
         if username != request.user.username and not request.user.is_admin:
-            return HttpResponseRedirect("/ClassRoom/Home")
+            return HttpResponseRedirect("/ClassRoom/"+classroom)
         file = Upload.objects.get(user=User.objects.get(username=username),
                                           quiz=Quiz.objects.get(pk=quiz_id),
                                           classroom=Quiz.objects.get(pk=quiz_id).classroom,
